Fixed implementation to show Sample data container
